+++ index.js
@@ -3,7 +3,7 @@
 const comments = require("./src/comments");
 const version = require("./package.json").version;
 const printAstToDoc = require("./src/printer").printAstToDoc;
-const getAlignmentSize = require("./src/printer").getAlignmentSize;
+const getAlignmentSize = require("./src/util").getAlignmentSize;
 const printDocToString = require("./src/doc-printer").printDocToString;
 const normalizeOptions = require("./src/options").normalize;
 const parser = require("./src/parser");
