+++ src/printer.js
@@ -49,25 +49,6 @@ function shouldPrintComma(options, level) {
   }
 }
 
-function getAlignmentSize(value, tabWidth, startIndex) {
-  startIndex = startIndex || 0;
-
-  let size = 0;
-  for (let i = startIndex; i < value.length; ++i) {
-    if (value[i] === '\t') {
-      // Tabs behave in a way that they are aligned to the nearest
-      // multiple of tabWidth:
-      // 0 -> 4, 1 -> 4, 2 -> 4, 3 -> 4
-      // 4 -> 8, 5 -> 8, 6 -> 8, 7 -> 8 ...
-      size = size + tabWidth - size % tabWidth;
-    } else {
-      size++;
-    }
-  }
-
-  return size;
-}
-
 function genericPrint(path, options, printPath, args) {
   assert.ok(path instanceof FastPath);
 
@@ -1684,7 +1665,7 @@ function genericPrintNoParens(path, options, print, args) {
           const index = value.lastIndexOf('\n');
           const tabWidth = options.tabWidth;
           if (index !== -1) {
-            size = getAlignmentSize(value, tabWidth, index + 1);
+            size = util.getAlignmentSize(value, tabWidth, index + 1);
           }
 
           let aligned = addAlignmentToDoc(expressions[i], size, tabWidth);
@@ -4287,4 +4268,4 @@ function addAlignmentToDoc(doc, size, tabWidth) {
   return aligned;
 }
 
-module.exports = { printAstToDoc, getAlignmentSize };
+module.exports = { printAstToDoc };
